Add a shell script to augment existing initrd
The directions in
Documentation/filesystems/ramfs-rootfs-initramfs.txt no longer
seem to work for creating a new initrd from an old one: only the
files in the kernel/ directory are added.  Provide a new shell
script that takes two arguments and adds a file to a subdirectoy
of a new initrd.  Tested on Debian Buster.